Move tests too; rename postings_compat.go
