+++ tsdb/index/postings_test.go
@@ -1101,117 +1101,6 @@ func BenchmarkMemPostings_Delete(b *testing.B) {
 	}
 }
 
-func TestPostingsCloner(t *testing.T) {
-	for _, tc := range []struct {
-		name  string
-		check func(testing.TB, *PostingsCloner)
-	}{
-		{
-			name: "seek beyond highest value of postings, then other clone seeks higher",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p1 := pc.Clone(context.TODO())
-				require.False(t, p1.Seek(9))
-
-				p2 := pc.Clone(context.TODO())
-				require.False(t, p2.Seek(10))
-			},
-		},
-		{
-			name: "seek beyond highest value of postings, then other clone seeks lower",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p1 := pc.Clone(context.TODO())
-				require.False(t, p1.Seek(9))
-
-				p2 := pc.Clone(context.TODO())
-				require.True(t, p2.Seek(2))
-				require.Equal(t, storage.SeriesRef(2), p2.At())
-			},
-		},
-		{
-			name: "seek to posting with value 3 or higher",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p := pc.Clone(context.TODO())
-				require.True(t, p.Seek(3))
-				require.Equal(t, storage.SeriesRef(4), p.At())
-				require.True(t, p.Seek(4))
-				require.Equal(t, storage.SeriesRef(4), p.At())
-			},
-		},
-		{
-			name: "seek alternatively on different postings",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p1 := pc.Clone(context.TODO())
-				require.True(t, p1.Seek(1))
-				require.Equal(t, storage.SeriesRef(1), p1.At())
-
-				p2 := pc.Clone(context.TODO())
-				require.True(t, p2.Seek(2))
-				require.Equal(t, storage.SeriesRef(2), p2.At())
-
-				p3 := pc.Clone(context.TODO())
-				require.True(t, p3.Seek(4))
-				require.Equal(t, storage.SeriesRef(4), p3.At())
-
-				p4 := pc.Clone(context.TODO())
-				require.True(t, p4.Seek(5))
-				require.Equal(t, storage.SeriesRef(8), p4.At())
-
-				require.True(t, p1.Seek(3))
-				require.Equal(t, storage.SeriesRef(4), p1.At())
-				require.True(t, p1.Seek(4))
-				require.Equal(t, storage.SeriesRef(4), p1.At())
-			},
-		},
-		{
-			name: "iterate through the postings",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p1 := pc.Clone(context.TODO())
-				p2 := pc.Clone(context.TODO())
-
-				// both one step
-				require.True(t, p1.Next())
-				require.Equal(t, storage.SeriesRef(1), p1.At())
-				require.True(t, p2.Next())
-				require.Equal(t, storage.SeriesRef(1), p2.At())
-
-				require.True(t, p1.Next())
-				require.Equal(t, storage.SeriesRef(2), p1.At())
-				require.True(t, p1.Next())
-				require.Equal(t, storage.SeriesRef(4), p1.At())
-				require.True(t, p1.Next())
-				require.Equal(t, storage.SeriesRef(8), p1.At())
-				require.False(t, p1.Next())
-
-				require.True(t, p2.Next())
-				require.Equal(t, storage.SeriesRef(2), p2.At())
-				require.True(t, p2.Next())
-				require.Equal(t, storage.SeriesRef(4), p2.At())
-			},
-		},
-		{
-			name: "ensure a failed seek doesn't allow more next calls",
-			check: func(t testing.TB, pc *PostingsCloner) {
-				p := pc.Clone(context.TODO())
-				require.False(t, p.Seek(9))
-				require.False(t, p.Next())
-			},
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			pc := NewPostingsCloner(newListPostings(1, 2, 4, 8))
-			tc.check(t, pc)
-		})
-	}
-
-	t.Run("cloning an err postings", func(t *testing.T) {
-		expectedErr := errors.New("foobar")
-		pc := NewPostingsCloner(ErrPostings(expectedErr))
-		p := pc.Clone(context.TODO())
-		require.False(t, p.Next())
-		require.Equal(t, expectedErr, p.Err())
-	})
-}
-
 func TestFindIntersectingPostings(t *testing.T) {
 	t.Run("multiple intersections", func(t *testing.T) {
 		p := NewListPostings([]storage.SeriesRef{10, 15, 20, 25, 30, 35, 40, 45, 50})
